src/display/requirements.txt
# Web Server
cherrypy
jinja2

# App
apscheduler

# Screen control
Pillow

# if /sys/bus/platform/drivers/gpiomem-bcm2835:
RPi.GPIO
spidev

# Sonarr Data
requests
